home.js, home.css, app.js, app.css, whatsappfloatingbutton.js
changed "learn more.." to view events. changed categories to Events.  Added a new js file called whatsappfloatingbutton.js to create a floating whatsapp button to the bottom right of the screen that is fixed for all the screen, and installed a new package using "npm install @fortawesome/fontawesome-free" to create the whatsapp logo. added an animation to the whatsapp button in app.css, as well as in home.css (pulsing effect)